fix codecheck

+++ common/include/interfaces/trust_device_profile.h
@@ -56,8 +56,8 @@ public:
     void SetPeerAccountId(const std::string& peerAccountId);
     std::string GetLocalAccountId() const;
     void SetLocalAccountId(const std::string& localAccountId);
-    std::vector<int32_t> GetServiceIdList();
-    void SetServiceIdList(const std::vector<int32_t> serviceIdList);
+    std::vector<int64_t> GetServiceIdList();
+    void SetServiceIdList(const std::vector<int64_t> serviceIdList);
     bool Marshalling(MessageParcel& parcel) const override;
     bool UnMarshalling(MessageParcel& parcel) override;
     std::string dump() const override;
